fix(review/1): move from internal to pkg
