+++ cmd/linode-cosi-driver/main.go
@@ -28,10 +28,10 @@ import (
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/logging"
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
-	"github.com/linode/linode-cosi-driver/internal/endpoint"
-	"github.com/linode/linode-cosi-driver/internal/envflag"
-	"github.com/linode/linode-cosi-driver/internal/grpc/handlers"
-	"github.com/linode/linode-cosi-driver/internal/grpc/logger"
+	"github.com/linode/linode-cosi-driver/pkg/endpoint"
+	"github.com/linode/linode-cosi-driver/pkg/envflag"
+	"github.com/linode/linode-cosi-driver/pkg/grpc/handlers"
+	"github.com/linode/linode-cosi-driver/pkg/grpc/logger"
 	"github.com/linode/linode-cosi-driver/pkg/servers/identity"
 	"github.com/linode/linode-cosi-driver/pkg/servers/provisioner"
 	"google.golang.org/grpc"
